Add next-cmake build system (#1985)
This is a parallel build tree that is being written to address structural issues in the main build which cannot be resolved incrementally (most notably the war that the python even interacts with cmake and the library build). Since it is parallel, we will complete it in several followon steps vs all at once.

---------

Co-authored-by: David Dixon <david.dixon@amd.com>